requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
mysqlclient==2.2.4
pillow==11.0.0
platformdirs==4.3.6
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.9.0
sqlparse==0.5.1
tzdata==2024.2
virtualenv==20.26.6

